+++ ICD.Connect.UI/Utils/KeypadStringBuilder.cs
@@ -13,6 +13,7 @@ namespace ICD.Connect.UI.Utils
 		/// <summary>
 		/// Raised when the string value changes.
 		/// </summary>
+		[PublicAPI]
 		public event EventHandler<StringEventArgs> OnStringChanged;
 
 		[NotNull]
@@ -32,7 +33,19 @@ namespace ICD.Connect.UI.Utils
 		/// Sets the current string in the builder.
 		/// </summary>
 		/// <param name="value"></param>
+		[PublicAPI]
 		public void SetString(string value)
+		{
+			SetString(value, false);
+		}
+
+		/// <summary>
+		/// Sets the current string in the builder.
+		/// </summary>
+		/// <param name="value"></param>
+		/// <param name="suppressEvent">If true, OnStringChanges event will be suppressed - Use with care</param>
+		[PublicAPI]
+		public void SetString(string value, bool suppressEvent)
 		{
 			value = value ?? string.Empty;
 			if (value == m_Output)
@@ -40,13 +53,15 @@ namespace ICD.Connect.UI.Utils
 
 			m_Output = value;
 
-			OnStringChanged.Raise(this, new StringEventArgs(m_Output));
+			if(!suppressEvent)
+				OnStringChanged.Raise(this, new StringEventArgs(m_Output));
 		}
 
 		/// <summary>
 		/// Appends the given character to the string.
 		/// </summary>
 		/// <param name="character"></param>
+		[PublicAPI]
 		public void AppendCharacter(char character)
 		{
 			SetString(m_Output + character);
@@ -55,6 +70,7 @@ namespace ICD.Connect.UI.Utils
 		/// <summary>
 		/// Removes the last character in the string.
 		/// </summary>
+		[PublicAPI]
 		public void Backspace()
 		{
 			if (!string.IsNullOrEmpty(m_Output))
@@ -64,15 +80,27 @@ namespace ICD.Connect.UI.Utils
 		/// <summary>
 		/// Clears the string.
 		/// </summary>
+		[PublicAPI]
 		public void Clear()
 		{
-			SetString(string.Empty);
+			Clear(false);
+		}
+
+		/// <summary>
+		/// Clears the string
+		/// </summary>
+		/// <param name="suppressEvent">If true, OnStringChanges event will be suppressed - Use with care</param>
+		[PublicAPI]
+		public void Clear(bool suppressEvent)
+		{
+			SetString(string.Empty, suppressEvent);
 		}
 
 		/// <summary>
 		/// Clears the builder and returns the contents.
 		/// </summary>
 		/// <returns></returns>
+		[PublicAPI]
 		[NotNull]
 		public string Pop()
 		{
@@ -85,6 +113,7 @@ namespace ICD.Connect.UI.Utils
 		/// Gets the resulting string.
 		/// </summary>
 		/// <returns></returns>
+		[PublicAPI]
 		[NotNull]
 		public override string ToString()
 		{